Create services file to handle service registering
Former configuration was replacing the services in `remote` and breaking all other remotes as the implementation was not the expected one and some services were missing.

This configures services per entity and only for the domain. This has some I/O on the event loop that will be handled in an upcoming change.